final commit

+++ address-book-frontend/src/app/pages/home/home.component.ts
@@ -1,5 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
 import { AddressBookService } from '../../services/address-book.service';
+import { AuthService } from '../../services/auth.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-home',
@@ -9,7 +11,12 @@ import { AddressBookService } from '../../services/address-book.service';
 export class HomeComponent implements OnInit {
   addresses: any[] = [];  
 
-  constructor(private addressBookService: AddressBookService) {}
+  constructor(
+    private addressBookService: AddressBookService,
+    private router: Router,
+    private authService:AuthService,
+    private cdr: ChangeDetectorRef 
+  ) {}
 
   ngOnInit(): void {
     this.fetchAddresses();
@@ -20,27 +27,31 @@ export class HomeComponent implements OnInit {
       next: (response) => {
         console.log("✅ Addresses fetched:", response);
         this.addresses = response;
+        this.cdr.detectChanges(); 
       },
       error: (err) => {
         console.error('❌ Error fetching addresses:', err);
       }
     });
   }
-
   deleteAddress(id: number): void {
-   
+    const confirmDelete = confirm('Are you sure you want to delete this address?');
+    if (!confirmDelete) return;
   
     this.addressBookService.deleteAddress(id).subscribe({
       next: () => {
-        alert("✅ Address deleted successfully!");
-        console.log("✅ Address deleted successfully!");
+        console.log('🗑️ Address deleted successfully');
         this.fetchAddresses(); 
       },
-      error: (error) => {
-        console.error("❌ Error deleting address:", error);
-        alert("❌ Error deleting address. Please try again.");
+      error: (err) => {
+        console.error('❌ Error deleting address:', err);
       }
     });
   }
-  
+
+  logout(): void {
+    this.authService.logout();
+    this.router.navigate(['auth/login']);
+  }
 }
+
